requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.2
discord.py==1.7.3
greenlet==1.1.2
idna==3.3
multidict==6.0.2
protobuf==3.19.4
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
twilio==7.6.0
twitchAPI==2.5.2
typing_extensions==4.1.1
urllib3==1.26.8
websockets==10.1
yarl==1.7.2